New Ship: Modified Arrow
This is more of a quick test to see if I got the whole branch thing down right.

This adds a new ship, the Modified Arrow, to the northern pirate shipyards! This arrow has been modified to be a really quick cargo ship, sporting 70 cargo with two guns and one turret.